refactor: move `streamText` to `text.ts`
