+++ src/helper/streaming/index.ts
@@ -1,21 +1,6 @@
-import { Context, HeaderRecord, TEXT_PLAIN } from '../../context'
-import { StatusCode } from '../../utils/http-status'
+import type { Context, HeaderRecord } from '../../context'
+import type { StatusCode } from '../../utils/http-status'
 import { StreamingApi } from '../../utils/stream'
-
-export { streamSSE } from './sse'
-
-export const streamText = (
-  c: Context,
-  cb: (stream: StreamingApi) => Promise<void>,
-  arg?: StatusCode | ResponseInit,
-  headers?: HeaderRecord
-): Response => {
-  headers ??= {}
-  c.header('content-type', TEXT_PLAIN)
-  c.header('x-content-type-options', 'nosniff')
-  c.header('transfer-encoding', 'chunked')
-  return stream(c, cb, arg, headers)
-}
 
 export const stream = (
   c: Context,
@@ -31,3 +16,6 @@ export const stream = (
     ? c.newResponse(readable, arg, headers)
     : c.newResponse(readable, arg)
 }
+
+export { streamSSE } from './sse'
+export { streamText } from './text'
